A) Clean up a bunch of duplication and centralize the way location settings are handled.
This allows us to have multiple user locations, and simplifies the API for associating locations with other things.

The variables used by location got all shuffled around because of this, but I put in compatibility code to keep setting location_maxnum_{nodetype} and location_defaultnum_{nodetype}, because there is a lot of contrib modules checking for them.

B) Made user locations work more like node locations.

C) Allow more than one location to be "required".

D) Reorganize the variables. Now there is a single variable per content type. (location_settings_{nodetype})

E) Make virtual fields hideable, and add the map link as a virtual field.

F) API change: The return value of 'virtual fields' op is now an array of keys and values, where the values are the labels for the hide checkboxes.

(The Drupal 5 version of this is a quick as possible port, there may be lingering typos. It seemed to work OK, but I won't know for sure until I minimize the diff against HEAD again.)