#342163 by jgadrow, update routine by me: Add missing prefix to the db_next_id() call.
If someone manages to hit this with postgresql, I'd be quite suprised, as the call has been broken since March and it is rather obvious when db_next_id is called against something invalid on postgresql -- it throws an error.

I put a message in the postgresql case to followup on the issue.